requirements.txt
polars==1.30.0      
pandas==2.3.0
numpy==2.3.0
matplotlib==3.10.3
statsmodels==0.14.4
scikit-learn==1.7.0
pyarrow==20.0.0             
aiohttp==3.12.11                
aiofiles==24.1.0              
tqdm==4.67.1               
hydra-core==1.3.2          
ipykernel==6.29.5
pytest==8.4.0
gymnasium==1.1.1
zarr==3.0.8
torch==2.7.1